% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_implemented_vigenttes.R
\name{add_rule_vignette}
\alias{add_rule_vignette}
\title{Create a data quality summary report based on the supplied 'rules'.
Current implementation gives:}
\usage{
add_rule_vignette(data, .rules, data_report, name = NULL, ...)
}
\arguments{
\item{data}{A data frame where the columns are the variable to be summarized.}

\item{.rules}{A list of `dq_rules` to be applied}

\item{data_report}{An existing `DataReport` object to add the report to}

\item{name}{The name for the file/ report}

\item{...}{Other parameters to be passed to the automated report e.g. 'description'}
}
\description{
TODO: Write this
}
\examples{
\dontrun{
TODO: write this
}

}
