% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dq_rules.R
\name{apply_rule}
\alias{apply_rule}
\title{Apply 'data quality rules' to a given data set.}
\usage{
apply_rule(data, rule)
}
\description{
Apply 'data quality rules' to a given data set.
}
\examples{
.rule <- new_rule("Example", function(i) !is.na(i))
apply_rule(iris, .rule)

}
