% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_bi_plot.R
\name{bi.plot}
\alias{bi.plot}
\title{General dispatch method for a bi-plot, intended for use
on a 'prcomp' object.}
\usage{
bi.plot(pca, x.index, y.index)
}
\arguments{
\item{pca}{A `prcomp` object}

\item{x.index}{numeric index of principal component to plot.}

\item{y.index}{numeric index of principal component to plot.}
}
\description{
General dispatch method for a bi-plot, intended for use
on a 'prcomp' object.
}
\examples{
mat <- model.matrix(~., data=iris)
pca <- prcomp(mat)
bi.plot(pca, 1, 2)
}
