% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extract.R
\name{feature_extract}
\alias{feature_extract}
\title{Define summary features for passing to data report template.}
\usage{
feature_extract(data)
}
\arguments{
\item{data}{The data set to be summarized}
}
\description{
Define summary features for passing to data report template.
}
\examples{
feature_extract(iris)
}
