% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extract.R
\name{group_standardize}
\alias{group_standardize}
\title{Correct scaling for categorical variables.}
\usage{
group_standardize(data, cols)
}
\description{
Standardizes so that groups of variables from the same factor have total var
of one, and each variable has zero-mean.
}
\examples{
mat <- as_tibble(model.matrix(~ Species, data=iris))
group_standardize(mat, c("Speciesversicolor", "Speciesvirginica"))
}
