% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extract.R
\name{is.binary}
\alias{is.binary}
\title{Check if a vector is binary (only has values 1, 0, TRUE, or FALSE)}
\usage{
is.binary(vec)
}
\description{
Check if a vector is binary (only has values 1, 0, TRUE, or FALSE)
}
\examples{

vec <- sample(c(0,1), 30, T)
is.binary(vec) # True

vec <- sample(c(0,1, T, F), 30, T)
is.binary(vec) # True

vec <- sample(4, 30, T)
is.binary(vec) # False

}
