% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_report.R
\name{new_data_report}
\alias{new_data_report}
\title{Creates a data report project structure and object.}
\usage{
new_data_report(path = NULL)
}
\arguments{
\item{path}{The path to the data report root}
}
\description{
Creates a data report project structure and object.
}
\examples{
\dontrun{
  new_data_report('~/Example.Data.Report')
}

}
