% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dq_rules.R
\name{new_rule}
\alias{new_rule}
\title{Create a 'data quality rule' for value level measures.}
\usage{
new_rule(name, .rule, .sel = everything())
}
\description{
For consistency with other code in the package we recommend writing rules
to identify the desired behavior i.e. 'True'  reflecting a pass and 'false'
a failure.
}
\examples{
new_rule("Example rule", function(vec) !is.na(vec))

}
