% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dq_rules.R
\name{new_variable_rule}
\alias{new_variable_rule}
\title{Create a 'data quality rule' for variable level measures.}
\usage{
new_variable_rule(...)
}
\description{
Create a 'data quality rule' for variable level measures.
}
\examples{
new_variable_rule("Example rule", function(vec) mean(is.na(vec)) < 0.1)
}
