% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsed_ratios.R
\name{parsed_ratios}
\alias{parsed_ratios}
\title{Summarize how much of each variable fits within a defined data type.}
\usage{
parsed_ratios(object)
}
\description{
Summarize how much of each variable fits within a defined data type.
}
\examples{
parsed_ratios(iris)

}
