% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extract.R
\name{styled_model_matrix}
\alias{styled_model_matrix}
\title{Produce a model.matrix embedding with variables named as defined by `style_func`}
\usage{
styled_model_matrix(data, formula = ~., style_func = style_contrasts)
}
\description{
Produce a model.matrix embedding with variables named as defined by `style_func`
}
\examples{
styled_model_matrix(iris, Sepal.Length ~ Sepal.Width + Species)

}
