% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styled_reports.R
\name{styled_variable_report}
\alias{styled_variable_report}
\title{Apply `datatable` styling to the variable rules from a `dq_rule_report`}
\usage{
styled_variable_report(report)
}
\arguments{
\item{report}{the output of `dq_rule_report`}
}
\description{
Apply `datatable` styling to the variable rules from a `dq_rule_report`
}
\examples{
rule <- new_rule("A", function(vec) !is.na(vec))
variable_rule <- new_variable_rule("A", function(vec) mean(is.na(vec)) < 0.1)

applied.rules <- apply_rules(iris, rule, variable_rule)

report <- dq_rule_report(applied.rules)

styled_variable_report(report)


}
