% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed_data_frame.R
\name{typed_summary_metrics}
\alias{typed_summary_metrics}
\title{Produce a list of data-type specific variable summaries.}
\usage{
typed_summary_metrics(data, func = summarize_typed_frame)
}
\description{
Produce a list of data-type specific variable summaries.
}
\examples{
typed_summary_metrics(iris)

}
