% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_there.R
\name{use_there}
\alias{use_there}
\title{Allow for `usethis` package functions to be evaluated within the data
report package}
\usage{
use_there(path, ...)
}
\description{
Allow for `usethis` package functions to be evaluated within the data
report package
}
